refactored modularity by removing unecessary folders, added first test for Chart component
