+++ frontend-engineer/frontend-assessment-app/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import React from 'react';
-import Chart from './components/Chart/Chart'
-import Map from './components/Map/Map'
+import Chart from './components/Chart'
+import Map from './components/Map'
 
 function App() {
   return (
