+++ frontend-engineer/frontend-assessment-app/src/components/Chart.tsx
@@ -3,7 +3,7 @@ import Highcharts from 'highcharts'
 import HighchartsReact from 'highcharts-react-official';
 import Exporting from 'highcharts/modules/exporting';
 import ExportData from 'highcharts/modules/export-data'
-import { useDataLoader } from '../../hooks/useDataLoader';
+import { useDataLoader } from '../hooks/useDataLoader';
 import Accessibility from 'highcharts/modules/accessibility'
 
 if(typeof Exporting === 'function'){
