+++ frontend-engineer/frontend-assessment-app/src/components/Map.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import mapboxgl from 'mapbox-gl';
 import 'mapbox-gl/dist/mapbox-gl.css'
-import { useDataLoader } from '../../hooks/useDataLoader';
+import { useDataLoader } from '../hooks/useDataLoader';
 
 
 mapboxgl.accessToken = 'pk.eyJ1IjoiYmxhc3Q2MSIsImEiOiJjbHltN21scnowdG95MmtwcnB6Yjd2dDJhIn0.PM7izJTLpphLWcN2VJvzPA'
